Fix: Change rawSize type from uint64_t to qint64 for type consistency
Changed the type of rawSize variable from uint64_t to qint64 to match the return type of QFile::size().

+++ voreen/src/qt/widgets/volumeiohelper.cpp
@@ -463,7 +463,7 @@ void VolumeIOHelper::loadRawVolume(const std::string& filenameStd) {
     else if (objectModel.find("TENSOR_") == 0)
         numChannels = 6;
 
-    uint64_t rawSize = headerSkip + static_cast<uint64_t>(formatBytes) * static_cast<uint64_t>(numChannels) * static_cast<uint64_t>(dim.x) * static_cast<uint64_t>(dim.y) * static_cast<uint64_t>(dim.z) * static_cast<uint64_t>(numFrames);
+    qint64 rawSize = headerSkip + formatBytes * numChannels * static_cast<qint64>(dim.x) * static_cast<qint64>(dim.y) * static_cast<qint64>(dim.z) * numFrames;
 
     // inform/query user, if file size does not match
     if (QFile(filename).size() != rawSize) {
